% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{extract_HFD_items}
\alias{extract_HFD_items}
\title{Extract names of HFD items from HTML}
\usage{
extract_HFD_items(html_content, CNTRY)
}
\arguments{
\item{html_content}{An object of class \code{xml_document} taken from a
country page of the HFD website, typically the result of
\code{httr::content(httr::GET(url))}}

\item{CNTRY}{A string given one of the country codes used by HFD. See
\code{getHFDcountries} for a function which lists available countries.}
}
\value{
A vector of characters, each of which is an item available for
download
}
\description{
Extract names of Human fertility database datasets from an
HTML page. Links are extracted from html, and for each link, any text falling
between the provided country code and a .txt extension is returned
}
