% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HMDutils.R
\name{getCHMDprovinces}
\alias{getCHMDprovinces}
\title{get a named vector of CHMD province codes}
\usage{
getCHMDprovinces()
}
\value{
a character vector of 3 character province codes.
}
\description{
This is a helper function to get a vector of 3-character province codes.
}
\examples{
\dontrun{ (provs <- getCHMDprovinces()) }
}

