% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{getHFCcountries}
\alias{getHFCcountries}
\title{getHFCcountries a function to grab all present country codes used in the Human Fertility Collection}
\usage{
getHFCcountries(names = FALSE)
}
\arguments{
\item{names}{logical. Default \code{FALSE} Should a \code{data.frame} matching full country names to short codes be given?}
}
\value{
either a character vector of short codes (default) or a \code{data.frame} of country names and codes.
}
\description{
The function returns a list of population codes used in the Human Fertiltiy Collection (HFC). Optionally, it also can return a data.frame with both the full population name and short code.
}
\examples{
\dontrun{
getHFCcountries()
getHFCcountries(names = TRUE)
}
}
