% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{getHFDcountries}
\alias{getHFDcountries}
\title{internal function for grabbing the HFD country short codes.}
\usage{
getHFDcountries()
}
\value{
a `tibble` with three columns `Country`, `link` and `CNTRY` (the country short code)
}
\description{
This function is called by \code{readHFDweb()} and is separated here for modularity. We include both main and provisional countries in the grab.
}
