% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{getHFDdate}
\alias{getHFDdate}
\title{internal function for grabbing the date of last update for a given HFD country}
\usage{
getHFDdate(CNTRY)
}
\arguments{
\item{CNTRY}{HFD country short code.}
}
\value{
character string of eight integers representing the date as \code{"yyyymmdd"}.
}
\description{
called by \code{readHFDweb()}. This assumes that \code{CNTRY} is actually available in the HFD.
}

