% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HMDutils.R
\name{getHMDcountries}
\alias{getHMDcountries}
\title{internal function for grabbing the HMD country short codes.}
\usage{
getHMDcountries()
}
\value{
a vector of HMD country short codes.
}
\description{
This function is called by \code{readHMDweb()} and is separated here for modularity. Assumes you have an internet connection.
}

