% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDutils.R
\name{getHMDitemavail}
\alias{getHMDitemavail}
\title{internal function for grabbing the available data item names for a given country.}
\usage{
getHMDitemavail(CNTRY)
}
\arguments{
\item{CNTRY}{character. HMD country short code.}
}
\value{
a tibble of all available data items for the selected country. There are several useful identifiers that can help determine the appropriate file, including the `measure`, `lexis`, `sex` and interval information, as detected from the item names.
}
\description{
called by \code{readHMDweb()} to find file urls. This assumes that \code{CNTRY} is actually available in the HFD.
}
