% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HMDutils.R
\name{getJMDprefectures}
\alias{getJMDprefectures}
\title{get a named vector of JMD prefecture codes}
\usage{
getJMDprefectures()
}
\value{
a character vector of 2-digit prefecture codes. Names correspond to the proper names given in the English version of the HMD webpage.
}
\description{
This is a helper function for those familiar with prefecture names but not with prefecture codes (and vice versa). It is also useful for looped downloading of data.
}
\examples{
\dontrun{ (prefectures <- getJMDprefectures()) }
}

