% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{getHFDcountries}
\alias{getHFDcountries}
\title{internal function for grabbing the HFD country short codes.}
\usage{
getHFDcountries()
}
\value{
a vector of HFD country short codes.
}
\description{
This function is called by \code{readHFDweb()} and is separated here for modularity. There is likely a simpler way of coding this functionality. The vector of short codes returned only includes the fully incorporated HFD countries, not provisional countries.
}

