% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{getHFDitemavail}
\alias{getHFDitemavail}
\title{List the available data item names for a given HFD country.}
\usage{
getHFDitemavail(CNTRY)
}
\arguments{
\item{CNTRY}{HFD country short code.}
}
\value{
a tibble of all available data files for the selected country. There are several useful identifiers that can help determine the appropriate file, including the `measure` and `subtype` as detected from the html table properties, and `lexis` and `parity` as detected either from the file names or the table properties.
}
\description{
called by \code{readHFDweb()}. This assumes that \code{CNTRY} is actually available in the HFD.
}
