% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HFDutils.R
\name{getHFDitemavail}
\alias{getHFDitemavail}
\title{internal function for grabbing the available data item names for a given country.}
\usage{
getHFDitemavail(CNTRY)
}
\arguments{
\item{CNTRY}{HFD country short code.}
}
\value{
a vector of item names. These are the file base names, and only need the extension \code{.txt} added in order to get te file name.
}
\description{
called by \code{readHFDweb()}. This assumes that \code{CNTRY} is actually available in the HFD.
}

