% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDutils.R
\name{getHMDitemavail}
\alias{getHMDitemavail}
\title{internal function for grabbing the available data item names for a given country.}
\usage{
getHMDitemavail(CNTRY, username, password)
}
\arguments{
\item{CNTRY}{character. HMD country short code.}

\item{username}{character. Your HMD user id usually the email address you registered with the HMD under.}

\item{password}{character. Your HMD password.}
}
\value{
character vector of item names. These are the file base names, and only need the extension \code{.txt} added in order to get the file name.
}
\description{
called by \code{readHMDweb()}. This assumes that \code{CNTRY} is actually available in the HFD.
}
